requirements.txt
python-telegram-bot==20.3
git+https://github.com/xtekky/g4f.git
